Fix runtime linker auditing on aarch64 V3
The dynamic linker's auditing was not working on aarch64. There were two
distinct problems:
  * _dl_runtime_resolve was not preserving x8 the indirect result location
    register.
  * The NEON Q registers pushed onto the stack by _dl_runtime_resolve
    were twice the size of D registers extracted from the stack frame by
    _dl_runtime_profile.

To fix this
  * The La_aarch64_regs structure was expanded to include x8 and the full
    sized NEON V registers that are required to be preserved by the ABI.
  * _dl_runtime_profile needed to extract registers saved by
    _dl_runtime_resolve and put them into the new correctly sized
    La_aarch64_regs structure.
  * The return value structure La_aarch64_retval also didn't have the correctly
    sized NEON V registers.

As a couple of additional cleanups
  * The names of the NEON registers saved within the La_aarch64_regs and the
    La_aarch_retval structures referred to the old D registers which were
    doubles. Now the registers are quads and are called V for vector registers.
    So the name of the field in the structure and the names of the offsets
    within that structure were renamed to use the more modern names.
  * The ABI specification says that r0-r7 register as well as the NEON v0-v7
    registers can be used to return values from a function. Therefore, I addded
    those to the La_aarch64_retval structure so that it also correctly matches
    the ABI.

Since this changes glibc's audit interface ABI, and because users want
to also at some point audit SVE functions, and we only want to break
ABI once.  The structure was further expaneded to include the required
registers required by the SVE variant of the ABI. Since these SVE
registers overlap with the V NEON registers, they are in a union with
them. To allow audit libraries to handle the variable sized SVE
registers and to allow them to know if they should look at the NEON or
SVE registers, a field was added to indicate the length of the SVE
registers at the time of the call.

Differences between V3 and V3 of the patch:
  * Switch the V registers back to __uint128_t as requested
  * Add support for SVE registers.

Differences between V2 and V1 of the patch:
  * V registers are now typed "long double" rather than "__uint128_t" an
    internal only type.
  * x8 is no longer preserved in the La_aarch64_retval
  * Comment describing the stack frame in _dl_runtime_profile had incorrect
    offsets after the size of structures changed.